restructure codebase for modularity
