+++ src/components/news/NewsHeader/NewsHeader.tsx
@@ -1,0 +1,65 @@
+import {
+  TableCell,
+  TableSortLabel,
+  TableRow,
+  TableHead,
+} from '@material-ui/core';
+import NewsModel from '../../../api/models/NewsModel';
+import { Order } from '../../../types';
+import useNewsListStyles from '../NewsList/NewsList.style';
+import { headCells } from './NewsHeader.content';
+
+interface EnhancedTableProps {
+  classes: ReturnType<typeof useNewsListStyles>;
+  onRequestSort: (
+    event: React.MouseEvent<unknown>,
+    property: keyof NewsModel
+  ) => void;
+  order: Order;
+  orderBy: string;
+  rowCount: number;
+}
+
+export default function EnhancedTableHead(props: EnhancedTableProps) {
+  const { classes, order, orderBy, onRequestSort } = props;
+  const createSortHandler =
+    (property: keyof NewsModel) => (event: React.MouseEvent<unknown>) => {
+      onRequestSort(event, property);
+    };
+
+  return (
+    <TableHead>
+      <TableRow>
+        {headCells.map((headCell) => (
+          <TableCell
+            className={classes.tableHead}
+            key={headCell.id}
+            align={headCell.align}
+            padding={headCell.disablePadding ? 'none' : 'normal'}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            {headCell.id === 'date' ? (
+              <TableSortLabel
+                active={orderBy === headCell.id}
+                direction={orderBy === headCell.id ? order : 'asc'}
+                onClick={createSortHandler(headCell.id)}
+                className={classes.tableHead}
+              >
+                {headCell.label}
+                {orderBy === headCell.id ? (
+                  <span className={classes.visuallyHidden}>
+                    {order === 'desc'
+                      ? 'sorted descending'
+                      : 'sorted ascending'}
+                  </span>
+                ) : null}
+              </TableSortLabel>
+            ) : (
+              headCell.label
+            )}
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+}
